Allow DefineArray behavior mirror DefineList when handed undefined
With DefineList if you do: `canReflect.new(DefineList, undefined)` you
get an empty list. This makes DefineArray mirror that behavior.

Note that DefineList and DefineArray are different when called with
`new` like `new DefineArray(values)` and `new DefineList(values)`.
DefineArray behaves more like an Array.

The `can.new` symbol is where DefineArray behaves like DefineList. So
this is keeping with that. Closes #16